Lambdas migrations: More PR amendments
- Added `DryRunnable` command mix-in
- Use `__doc__` directive for help message
- Removed unnecessary `READWRITE` constant in one command
- Removed unnecessary imports
- Tweaked comment

NOTE: Inherit from `DryRunnable` doesn't mean it a command will not have
side effects. Your command will still have to add the logic to use the
`--dry-run` flag!
