+++ control_panel_api/management/commands/migrate_lambda_data_0_dump_user_policies.py
@@ -8,20 +8,18 @@ from django.core.management.base import BaseCommand
 from control_panel_api.models import User
 
 
-READWRITE = 'readwrite'
-
-
 logger = logging.getLogger(__name__)
 
 
 class Command(BaseCommand):
     """
-    NOTE: Needs permission to perform `iam:ListAttachedRolePolicies` action
+    For each of the users in the DB, it collects its attached IAM policies
+    and prints this data to stdout.
+
+    NOTE: Needs permission to perform `iam:ListAttachedRolePolicies` action.
     """
 
-
-    help = ("For each of the users in the DB, it collects its attached IAM "
-            "policies and prints this data to stdout")
+    help = __doc__
 
     def handle(self, *args, **options):
         iam = boto3.client('iam')
